Painters can now be used in a nested way for the upcoming contentOffset support.
The contentOffset support is not yet hooked up, as that will require a
bit more work (essentially, to surface the actual content offset to be
applied in the constructor to the nested painter from the scrollview),
but at least everything can now use the nested painting system and the
nested views work as expected.

This means that every redraw method now gets two parameters, the first
one is the region that must be redrawn (region:) in coordinates local
to the view, and then a painter that should be used in the redraw
method to draw, it is no longer necessary to request a painter
manually.

Containers that will render children objects directly must create a
new Painter object passing the parent painter.

Rect.interesection: fixed it so that empty rectangles are returned,
instead of returning negative sizes.

Rect.contains: off-by one error

TODO: currently the add() methods in Painter do not enforce clipping
on the column side, this needs to be added.   To make this more
obvious, I have removed the Application/driver clipping calls in
ScrollView.redraw, which make the problem more apparent when running
the sample.