Start implementation of toplevels with backbuffers, this currently is just a simple array of Cells (Character, Attribute) on the Toplevel.
I want to move all the code to use the new entry point in Painter
that takes a Toplevel for toplevel drawing.   This would render,
not via the driver, but would render into the backing store, which
we can later compose more easily.

For wide characters, this inserts the first character in the cell,
and then inserts a 0 character for the second cell, which should
have the same effect in terms of usage of the buffer as using
two cells, but the second is not updates.

The next steps are:

- Add code to copy the backingstore into the screen (that should
give me parity with the current code)

- Add code to detect dirty regions on the toplevels, and compose
the toplevels based on their backing stores.

- Ensure that we remove the Painter.createRootPainter, to force
all the users of the painter to either be the view painting which
has its clipping, or to draw into the backing store (we could
keep the name, but we need to ensure that nobody is using the
current entry point).

Once the above is done, it might be a good time to revisit
the postProcess event which is sort of a hack after an event
has been processed, and instead have a proper "Please update me"
event that is triggered when returning to the mainloop.