Fixed various rendering design flaws.
1. Eliminate "_childNeedsDisplay", instead, when a region in a view is
flagged for requiring an update, update the coordinates in the
container accordingly.  This means that we always have up to date
information across the entire hierarchy about what needs updates, and
not rely on just "a child has a damaged region", as we would not know
if this would impact other regions of the container view.

2. Fix a bug where "setNeedsDisplay(frame)" was called on self, instead
it needs to be "bounds" as the coordinates must be relative to the
receiver.

3. Fixes a bug in the computation of the needDisplay region, it was
using maxX, instead of the width, which was causing larger regions to
be painted, but most importanly, with the new framework that keeps
everything in sync it would trigger a stack overflow as the region kept
growing.

4. Because of the change where we call setNeedsDisplay() on the parent,
this means that the parent will call us back with the new region, so we
catch this new region update, and if it is the same, we do not continue
the regular processing, as we should now always be up to date.

5. Generally, calling 'drawFrame' with the "fill" set to true, will
paint over the child contents, which means that if they were not
expected to be redrawn, given that they are using needsDisplay region
to track the changes, they would not be repainted.   So it is important
that we do not clear regions that we are not asked to clear.

So p.clear() should not be used by containers, instead p.clear
(needDisplay) is a better idiom.