View/Toplevel: rather than manually triggering setNeedDisplay on focus changes, let the resignFirstResponder() and becomeFirstReponder() do the work.
setNeedsDisplay now properly accumulates the data for the modified region

It additionally, clips the region to the container, to not spill over
other views.

Since I dropped the message on become/release, simplified the method.

When computing the automatic layout, do not default the width of a
view to the width of the container, but to the remaining width.

View: improve debugDescription