Opens various types and variables, I keep using the C# idiom, forgetting that Swift does not treat public as open.
StandardTopLevel: a toplevel that has a menu, status bar and desktop,
and moved some of the capabilities that before were in Toplevel here
(like tracking active menu and statusbar).  Part of the redesign.

setNeedsLayout(): now propagates this requirement to the toplevel, and
is now treated and handled at the view level, and does not affect or
trigger a layout on the parent.

Application: Pipe wakeup system.  When a change that requires the
current postProcessEvent method to be called (layout or need display),
now we have an Application.requestPostProcess method that will wake up
the main thread and handle it.  Both setNeedsDisplay and
setNeedsLayout() not use this.

Application.requestStop: fixed, it was using the global queue to run
render(), which was wrong, it should have been main.   Also, use an
assertion, to make sure I am not doing something wrong.

Painter.clear: bug fix, this was not clearing the proper region, it
was doing the frame, not the bounds

Painter.clear: special case the code for Characers that have a single
UnicodeScalar

Painter.drawFrame: bug fix, when I made a window too small, this used
to crash.

New demo start: a text editor.