Make this build in modern systems.
The downside currently is that Packages.swift declares a minimum of MacOS 15.0,
which is merely necessary because Apple flagged ncurses this way.   Not true,
but have no patience to fight now, and will have a native provider at some
point that does not need it.

Drop OpenCombine, long term, I will write a TermKitUI stack that just leverages
Observable.